requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
Markdown==3.2.2
mysqlclient==2.0.1
PyJWT==1.7.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
djangorestframework-simplejwt
debugpy==1.6.6
importlib-metadata==4.13.0
stripe==5.3.0